requirements.txt
# Core Data Analysis
pandas==2.1.4
numpy==1.24.3
scipy==1.11.4

# Financial Data
yfinance==0.2.28
alpha-vantage==2.3.1
fredapi==0.5.1

# Technical Analysis
ta==0.10.2
talib-binary==0.4.26
mplfinance==0.12.10b0

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
xgboost==2.0.3

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
dash==2.16.1

# Web Framework
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Data Processing
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Utilities
python-dotenv==1.0.0
schedule==1.2.0
tqdm==4.66.1 